Reestruturação
Esta nova organização permite que que o pacote seja mais flexivel
permitindo a utilização de mais de um serviço/provedor de CEP
