+++ composer.json
@@ -20,7 +20,7 @@
   },
   "autoload": {
     "psr-4": {
-      "Canducci\\ZipCode\\": "src/Canducci/ZipCode"
+      "Canducci\\ZipCode\\": "src/Canducci/ZipCode/"
     },
     "files":[
       "src/Canducci/ZipCode/ZipCodeHelper.php"
