+++ src/Canducci/ZipCode/ZipCodeException.php
@@ -1,0 +1,6 @@
+<?php namespace Canducci\ZipCode;
+
+class ZipCodeException extends \Exception
+{
+
+}
