+++ src/Canducci/ZipCode/ZipCodeHelper.php
@@ -1,14 +1,13 @@
 <?php namespace {
     if (!function_exists('zipcode')) {
         /**
-         * @param $value
-         * @return $this
-         * @throws \Exception
+         * @param string $value
+         *
+         * @return \Canducci\ZipCode\Contracts\ZipCodeServiceContract
          */
         function zipcode($value)
         {
-            $zip_code = new Canducci\ZipCode\ZipCode();
-            return $zip_code->find($value);
+            return app('zipcode')->find($value);
         }
     }
 }
