+++ src/Canducci/config/zipcode.php
@@ -1,0 +1,5 @@
+<?php
+
+return [
+    'service' => 'Canducci\ZipCode\Services\ViacepService'
+];
